requirements.txt
requests
scapy
python-nmap